Create a specialButtons map and move EditButton and CloseButton into this map
