+++ frontend/src/specialButtons/CloseButton.js
